+++ src/main/kotlin/com/dobby/backend/domain/exception/ErrorCode.kt
@@ -46,11 +46,11 @@ enum class ErrorCode(
     /**
      * Signup error codes
      */
-    SIGNUP_ALREADY_MEMBER("SI001", "You've already joined", HttpStatus.CONFLICT),
+    SIGNUP_ALREADY_MEMBER("SIGN_UP_001", "You've already joined", HttpStatus.CONFLICT),
 
     /**
-     * Login error codes
+     * Signin error codes
      */
-    LOGIN_MEMBER_NOT_FOUND("LO001", "Member Not Found. Please signup. (Redirect URI must be /v1/auth/{role}/signup)", HttpStatus.NOT_FOUND),
-    LOGIN_ROLE_MISMATCH("LO002", "Already registered member as %s. Please login as %s", HttpStatus.BAD_REQUEST)
+    SIGNIN_MEMBER_NOT_FOUND("SIGN_IN_001", "Member Not Found. Please signup. (Redirect URI must be /v1/auth/{role}/signup)", HttpStatus.NOT_FOUND),
+    SIGNIN_ROLE_MISMATCH("SIGN_IN_002", "Already registered member as %s. Please login as %s", HttpStatus.BAD_REQUEST)
 }
