+++ src/main/kotlin/com/dobby/backend/domain/exception/SignInRoleMismatchException.kt
@@ -3,16 +3,16 @@ package com.dobby.backend.domain.exception
 import com.dobby.backend.infrastructure.database.entity.enum.RoleType
 import org.springframework.http.HttpStatus
 
-class LoginRoleMismatchException(
+class SignInRoleMismatchException(
     registeredRole: RoleType?,
     requestedRole: RoleType
 ) : RuntimeException(
     String.format(
-        ErrorCode.LOGIN_ROLE_MISMATCH.message,
+        ErrorCode.SIGNIN_ROLE_MISMATCH.message,
         registeredRole?.name,
         requestedRole.name
     )
 ) {
-    val code: String = ErrorCode.LOGIN_ROLE_MISMATCH.code
-    val status: HttpStatus = ErrorCode.LOGIN_ROLE_MISMATCH.httpStatus
+    val code: String = ErrorCode.SIGNIN_ROLE_MISMATCH.code
+    val status: HttpStatus = ErrorCode.SIGNIN_ROLE_MISMATCH.httpStatus
 }
